requirements.txt
librosa>=0.8.0
pydub>=0.25.0
numpy>=1.19.0
pandas>=1.2.0
tqdm>=4.60.0
PyYAML>=5.4.0
scipy>=1.6.0
scikit-learn>=0.24.0
wget>=3.2
requests>=2.25.0
